+++ x-pack/solutions/security/plugins/security_solution/public/common/lib/integrations/components/__mocks__/index.tsx
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+import React from 'react';
+
+export const SecurityIntegrations = () => <div data-test-subj="securityIntegrations" />;
