+++ x-pack/solutions/security/plugins/security_solution/public/common/lib/integrations/hooks/__mocks__/integration_context.tsx
@@ -1,0 +1,19 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { IntegrationContextValue } from '../integration_context';
+import { mockIntegrationContext } from './mocks';
+
+export const useIntegrationContext = (): IntegrationContextValue => {
+  return mockIntegrationContext();
+};
