+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/lib/telemetry.ts
@@ -7,6 +7,8 @@
 
 import { METRIC_TYPE, TELEMETRY_EVENT, track } from '../../../common/lib/telemetry';
 
+export type TrackLinkClick = (linkId: string) => void;
+
 export const trackOnboardingLinkClick = (linkId: string) => {
   track(METRIC_TYPE.CLICK, `${TELEMETRY_EVENT.ONBOARDING}_${linkId}`);
 };
