+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_body/cards/common/integrations/callouts/integration_card_top_callout.test.tsx
@@ -10,7 +10,7 @@ import { render, waitFor } from '@testing-library/react';
 import { of } from 'rxjs';
 import { IntegrationCardTopCallout } from './integration_card_top_callout';
 import { useOnboardingService } from '../../../../../hooks/use_onboarding_service';
-import { IntegrationTabId } from '../types';
+import { IntegrationTabId } from '../../../../../../../common/lib/integrations/types';
 
 jest.mock('../../../../../hooks/use_onboarding_service', () => ({
   useOnboardingService: jest.fn(),
