+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_body/cards/common/integrations/callouts/integration_card_top_callout.tsx
@@ -11,8 +11,8 @@ import useObservable from 'react-use/lib/useObservable';
 import { useOnboardingService } from '../../../../../hooks/use_onboarding_service';
 import { AgentlessAvailableCallout } from './agentless_available_callout';
 import { InstalledIntegrationsCallout } from './installed_integrations_callout';
-import { IntegrationTabId } from '../types';
 import { EndpointCallout } from './endpoint_callout';
+import { IntegrationTabId } from '../../../../../../../common/lib/integrations/types';
 
 export const useShowInstalledCallout = ({
   installedIntegrationsCount,
