+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_body/cards/common/integrations/callouts/manage_integrations_callout.test.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { ManageIntegrationsCallout } from './manage_integrations_callout';
 import { TestProviders } from '../../../../../../../common/mock/test_providers';
-
+jest.mock('../../../../../../../common/lib/integrations/hooks/integration_context');
 jest.mock('../../../../../../../common/hooks/use_add_integrations_url', () => ({
   useAddIntegrationsUrl: jest.fn().mockReturnValue({
     href: '/test-url',
