+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_body/cards/common/integrations/constants.ts
@@ -5,25 +5,11 @@
  * 2.0.
  */
 
-import type { CategoryFacet } from '@kbn/fleet-plugin/public';
-
 export const ADD_AGENT_PATH = `/agents`;
 export const AGENT_INDEX = `logs-elastic_agent*`;
-export const CARD_DESCRIPTION_LINE_CLAMP = 3; // 3 lines of text
-export const CARD_TITLE_LINE_CLAMP = 1; // 1 line of text
 export const FLEET_APP_ID = `fleet`;
-export const INTEGRATION_APP_ID = `integrations`;
-export const LOADING_SKELETON_TEXT_LINES = 10; // 10 lines of text
-export const MAX_CARD_HEIGHT_IN_PX = 127; // px
-export const RETURN_APP_ID = 'returnAppId';
-export const RETURN_PATH = 'returnPath';
-export const SCROLL_ELEMENT_ID = 'integrations-scroll-container';
-export const SEARCH_FILTER_CATEGORIES: CategoryFacet[] = [];
 export const INTEGRATION_CARD_HEIGHT = 156;
-export const DEFAULT_INTEGRATION_CARD_CONTENT_HEIGHT = `${INTEGRATION_CARD_HEIGHT * 3.5 + 55}px`;
 export const TELEMETRY_MANAGE_INTEGRATIONS = `manage_integrations`;
 export const TELEMETRY_ENDPOINT_LEARN_MORE = `endpoint_learn_more`;
 export const TELEMETRY_AGENTLESS_LEARN_MORE = `agentless_learn_more`;
 export const TELEMETRY_AGENT_REQUIRED = `agent_required`;
-export const TELEMETRY_INTEGRATION_CARD = `card`;
-export const TELEMETRY_INTEGRATION_TAB = `tab`;
