+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_body/cards/integrations/integrations_check_complete.ts
@@ -5,6 +5,7 @@
  * 2.0.
  */
 
+import type { IntegrationCardMetadata } from '../../../../../common/lib/integrations/types';
 import type { StartServices } from '../../../../../types';
 import type { OnboardingCardCheckComplete } from '../../../../types';
 import {
@@ -12,7 +13,6 @@ import {
   getCompleteBadgeText,
   getIntegrationList,
 } from '../common/integrations/integrations_check_complete';
-import type { IntegrationCardMetadata } from '../common/integrations/types';
 
 export const checkIntegrationsCardComplete: OnboardingCardCheckComplete<
   IntegrationCardMetadata
