+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_body/cards/search_ai_lake/integrations_search_ai_lake/index.ts
@@ -10,10 +10,10 @@ import { i18n } from '@kbn/i18n';
 import type { OnboardingCardConfig } from '../../../../../types';
 import { checkIntegrationsCardComplete } from './integrations_check_complete';
 import { OnboardingCardId } from '../../../../../constants';
-import type { IntegrationCardMetadata } from '../../common/integrations/types';
 import integrationsIcon from '../../common/integrations/images/integrations_icon.png';
 import integrationsDarkIcon from '../../common/integrations/images/integrations_icon_dark.png';
 import { SECURITY_FEATURE_ID } from '../../../../../../../common/constants';
+import type { IntegrationCardMetadata } from '../../../../../../common/lib/integrations/types';
 
 export const integrationsSearchAILakeCardConfig: OnboardingCardConfig<IntegrationCardMetadata> = {
   id: OnboardingCardId.integrationsSearchAILake,
