+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_body/cards/search_ai_lake/integrations_search_ai_lake/integrations_card.tsx
@@ -8,19 +8,20 @@ import React from 'react';
 
 import type { OnboardingCardComponent } from '../../../../../types';
 import { OnboardingCardContentPanel } from '../../common/card_content_panel';
-import { IntegrationsCardGridTabsComponent } from '../../common/integrations/integration_card_grid_tabs';
 import { CenteredLoadingSpinner } from '../../../../../../common/components/centered_loading_spinner';
-import type { IntegrationCardMetadata } from '../../common/integrations/types';
 import { INTEGRATION_TABS } from './integration_tabs_configs';
 import { ManageIntegrationsCallout } from '../../common/integrations/callouts/manage_integrations_callout';
-import {
-  WithFilteredIntegrations,
-  type RenderChildrenType,
-} from '../../common/integrations/use_filter_cards';
-import { useSelectedTab } from '../../common/integrations/use_selected_tab';
 import { useOnboardingContext } from '../../../../onboarding_context';
 import { useEnhancedIntegrationCards } from '../../../../../../common/lib/search_ai_lake/hooks';
-import { DEFAULT_CHECK_COMPLETE_METADATA } from '../../integrations/integrations_card';
+import { useSelectedTab } from '../../../../../../common/lib/integrations/hooks/use_selected_tab';
+import type {
+  RenderChildrenType,
+  IntegrationCardMetadata,
+} from '../../../../../../common/lib/integrations/types';
+import { WithFilteredIntegrations } from '../../../../../../common/lib/integrations/components/with_filtered_integrations';
+import { IntegrationsCardGridTabsComponent } from '../../../../../../common/lib/integrations/components/integration_card_grid_tabs_component';
+import { DEFAULT_CHECK_COMPLETE_METADATA } from '../../../../../../common/lib/integrations/components/integration_card_grid_tabs';
+import { IntegrationContextProvider } from '../../../../../../common/lib/integrations/hooks/integration_context';
 
 const IntegrationsCardGridTabs: RenderChildrenType = ({
   allowedIntegrations,
@@ -51,7 +52,10 @@ const IntegrationsCardGridTabs: RenderChildrenType = ({
 
 export const IntegrationsCard: OnboardingCardComponent<IntegrationCardMetadata> = React.memo(
   ({ checkCompleteMetadata }) => {
-    const { spaceId } = useOnboardingContext();
+    const {
+      spaceId,
+      telemetry: { trackLinkClick },
+    } = useOnboardingContext();
 
     const selectedTabResult = useSelectedTab({
       spaceId,
@@ -64,12 +68,14 @@ export const IntegrationsCard: OnboardingCardComponent<IntegrationCardMetadata> 
 
     return (
       <OnboardingCardContentPanel>
-        <WithFilteredIntegrations
-          renderChildren={IntegrationsCardGridTabs}
-          prereleaseIntegrationsEnabled={true}
-          checkCompleteMetadata={checkCompleteMetadata}
-          selectedTabResult={selectedTabResult}
-        />
+        <IntegrationContextProvider spaceId={spaceId} trackLinkClick={trackLinkClick}>
+          <WithFilteredIntegrations
+            renderChildren={IntegrationsCardGridTabs}
+            prereleaseIntegrationsEnabled={true}
+            checkCompleteMetadata={checkCompleteMetadata}
+            selectedTabResult={selectedTabResult}
+          />
+        </IntegrationContextProvider>
       </OnboardingCardContentPanel>
     );
   }
