+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_footer/onboarding_footer.test.tsx
@@ -7,11 +7,12 @@
 
 import React from 'react';
 import { render } from '@testing-library/react';
-import { trackOnboardingLinkClick } from '../lib/telemetry';
 import { FooterLinkItem } from './onboarding_footer';
 import { OnboardingFooterLinkItemId, TELEMETRY_FOOTER_LINK } from './constants';
+import { mockTrackLinkClick } from '../__mocks__/mocks';
 
 jest.mock('../lib/telemetry');
+jest.mock('../onboarding_context');
 
 describe('OnboardingFooterComponent', () => {
   beforeEach(() => {
@@ -45,7 +46,7 @@ describe('OnboardingFooterComponent', () => {
     );
 
     getByTestId('footerLinkItem').click();
-    expect(trackOnboardingLinkClick).toHaveBeenCalledWith(
+    expect(mockTrackLinkClick).toHaveBeenCalledWith(
       `${TELEMETRY_FOOTER_LINK}_${OnboardingFooterLinkItemId.documentation}`
     );
   });
