+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_footer/onboarding_footer.tsx
@@ -9,9 +9,9 @@ import React, { useCallback } from 'react';
 import { EuiFlexGroup, EuiFlexItem, EuiLink, EuiSpacer, EuiText, EuiTitle } from '@elastic/eui';
 import { useFooterStyles } from './onboarding_footer.styles';
 import { useFooterItems } from './footer_items';
-import { trackOnboardingLinkClick } from '../lib/telemetry';
 import type { OnboardingFooterLinkItemId } from './constants';
 import { TELEMETRY_FOOTER_LINK } from './constants';
+import { useOnboardingContext } from '../onboarding_context';
 
 export const OnboardingFooter = React.memo(() => {
   const styles = useFooterStyles();
@@ -43,9 +43,13 @@ interface FooterLinkItemProps {
 
 export const FooterLinkItem = React.memo<FooterLinkItemProps>(
   ({ id, title, icon, description, link }) => {
+    const {
+      telemetry: { trackLinkClick },
+    } = useOnboardingContext();
+
     const onClickWithReport = useCallback<React.MouseEventHandler>(() => {
-      trackOnboardingLinkClick(`${TELEMETRY_FOOTER_LINK}_${id}`);
-    }, [id]);
+      trackLinkClick?.(`${TELEMETRY_FOOTER_LINK}_${id}`);
+    }, [id, trackLinkClick]);
 
     return (
       <EuiFlexItem>
