+++ x-pack/solutions/security/plugins/security_solution/public/onboarding/components/onboarding_telemetry.ts
@@ -11,14 +11,20 @@ import type { OnboardingCardId } from '../constants';
 import { OnboardingTopicId } from '../constants';
 import { OnboardingHubEventTypes } from '../../common/lib/telemetry';
 import { onboardingConfig } from '../config';
+import type { TrackLinkClick } from './lib/telemetry';
 
 export interface OnboardingTelemetry {
   reportCardOpen: (cardId: OnboardingCardId, options?: { auto?: boolean }) => void;
   reportCardComplete: (cardId: OnboardingCardId, options?: { auto?: boolean }) => void;
   reportCardLinkClicked: (cardId: OnboardingCardId, linkId: string) => void;
+  trackLinkClick?: TrackLinkClick;
 }
 
-export const useOnboardingTelemetry = (): OnboardingTelemetry => {
+export const useOnboardingTelemetry = ({
+  trackLinkClick,
+}: {
+  trackLinkClick?: TrackLinkClick;
+}): OnboardingTelemetry => {
   const { telemetry } = useKibana().services;
   return useMemo(
     () => ({
@@ -40,8 +46,9 @@ export const useOnboardingTelemetry = (): OnboardingTelemetry => {
           stepLinkId: linkId,
         });
       },
+      trackLinkClick,
     }),
-    [telemetry]
+    [telemetry, trackLinkClick]
   );
 };
